requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-pinecone==0.2.0
pinecone-client==5.0.1
openai==1.51.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
tiktoken==0.8.0
httpx==0.27.2
